rename PelotonToGarminConsole to ConsoleClient
