+++ .github/workflows/sync_peloton_to_garmin.yml
@@ -73,7 +73,7 @@ jobs:
           }
         }
         EOT
-    - run: /app/PelotonToGarminConsole
+    - run: /app/ConsoleClient
       if: ${{ env.P2G_PELOTON__EMAIL && env.P2G_PELOTON__PASSWORD && env.P2G_GARMIN__EMAIL && env.P2G_GARMIN__PASSWORD }}
       working-directory: /app
       env:
