+++ src/ConsoleClient/256x256.ico
